Allow deleting a journal from logged users only

+++ controllers/journalController.js
@@ -76,7 +76,18 @@ journalsRouter.put('/:id', async (request, response) => {
 })
 
 journalsRouter.delete('/:id', async (request, response) => {
+	const token = getTokenFrom(request)
+
+	const decodedToken = jwt.verify(token, process.env.SECRET)
+	if (!token || !decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid' })
+	}
+	const user = await User.findById(decodedToken.id)
+
 	await Journal.findByIdAndRemove(request.params.id)
+	user.journals = user.journals.filter(journal => journal.id !== request.params.id)
+	await user.save()
+
 	response.status(204).end()
 })
